[Refactor] Split connector core interfaces from implementations
