+++ be/src/connector/connector_defaults.cpp
@@ -1,0 +1,41 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "connector/connector.h"
+
+#include "common/logging.h"
+
+namespace starrocks::connector {
+
+DataSourceProviderPtr Connector::create_data_source_provider(ConnectorScanNode* scan_node,
+                                                             const TPlanNode& plan_node) const {
+    CHECK(false) << connector_type() << " connector does not implement chunk source yet";
+    __builtin_unreachable();
+}
+
+std::unique_ptr<ConnectorChunkSinkProvider> Connector::create_data_sink_provider() const {
+    CHECK(false) << connector_type() << " connector does not implement chunk sink yet";
+    __builtin_unreachable();
+}
+
+std::unique_ptr<ConnectorChunkSinkProvider> Connector::create_delete_sink_provider() const {
+    CHECK(false) << connector_type() << " connector does not implement chunk sink yet";
+    __builtin_unreachable();
+}
+
+std::unique_ptr<ConnectorChunkSinkProvider> Connector::create_row_delta_sink_provider() const {
+    return nullptr;
+}
+
+} // namespace starrocks::connector
