+++ be/src/connector/connector_registry.cpp
@@ -1,0 +1,37 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "connector/connector_registry.h"
+
+#include "connector/connector.h"
+
+namespace starrocks::connector {
+
+static ConnectorRegistry _global_default_instance;
+
+const Connector* ConnectorRegistry::get(const std::string& name) {
+    auto it = _connectors.find(name);
+    if (it == _connectors.end()) return nullptr;
+    return it->second.get();
+}
+
+void ConnectorRegistry::put(const std::string& name, std::unique_ptr<Connector> connector) {
+    _connectors.emplace(name, std::move(connector));
+}
+
+ConnectorRegistry* ConnectorRegistry::default_instance() {
+    return &_global_default_instance;
+}
+
+} // namespace starrocks::connector
