+++ be/src/connector/connector_type.h
@@ -1,0 +1,32 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+namespace starrocks::connector {
+
+enum ConnectorType {
+    HIVE = 0,
+    ES = 1,
+    JDBC = 2,
+    MYSQL = 3,
+    FILE = 4,
+    LAKE = 5,
+    BINLOG = 6,
+    ICEBERG = 7,
+    BENCHMARK = 8,
+    CACHE_STATS = 9,
+};
+
+} // namespace starrocks::connector
