+++ be/src/exec/data_sinks/hive_table_sink.cpp
@@ -109,7 +109,7 @@ Status HiveTableSink::decompose_to_pipeline(pipeline::OpFactories prev_operators
     }
     sink_ctx->fragment_context = fragment_ctx;
 
-    auto connector = connector::ConnectorManager::default_instance()->get(connector::Connector::HIVE);
+    auto connector = connector::ConnectorRegistry::default_instance()->get(connector::Connector::HIVE);
     auto sink_provider = connector->create_data_sink_provider();
     auto op = std::make_shared<pipeline::ConnectorSinkOperatorFactory>(
             context->next_operator_id(), std::move(sink_provider), sink_ctx, fragment_ctx);
