+++ be/src/exec/data_sinks/iceberg_table_sink.cpp
@@ -245,7 +245,7 @@ Status IcebergTableSink::create_delete_sink_context(
     }
 
     sink_ctx = delete_sink_ctx;
-    auto connector = connector::ConnectorManager::default_instance()->get(connector::Connector::ICEBERG);
+    auto connector = connector::ConnectorRegistry::default_instance()->get(connector::Connector::ICEBERG);
     sink_provider = connector->create_delete_sink_provider();
 
     return Status::OK();
@@ -323,7 +323,7 @@ Status IcebergTableSink::create_data_sink_context(const TDataSink& thrift_sink, 
     }
 
     sink_ctx = data_sink_ctx;
-    auto connector = connector::ConnectorManager::default_instance()->get(connector::Connector::ICEBERG);
+    auto connector = connector::ConnectorRegistry::default_instance()->get(connector::Connector::ICEBERG);
     sink_provider = connector->create_data_sink_provider();
 
     if (iceberg_table_desc->is_unpartitioned_table()) {
@@ -561,7 +561,7 @@ Status IcebergTableSink::create_row_delta_sink_context(
     row_delta_ctx->op_code_index = op_code_index;
 
     sink_ctx = row_delta_ctx;
-    auto connector = connector::ConnectorManager::default_instance()->get(connector::Connector::ICEBERG);
+    auto connector = connector::ConnectorRegistry::default_instance()->get(connector::Connector::ICEBERG);
     sink_provider = connector->create_row_delta_sink_provider();
 
     return Status::OK();
